image/requirements.txt
pymodbus==2.5.3
opcua==0.98.13
PyYAML==6.0.3
python-nmap==0.7.1
Flask==3.1.2
requests==2.32.5